refactor(recorder): inline RecorderLifecycleState into LifecycleController, flatten lifecycle/ folder
