+++ extension/src/extension/services/telemetry/recording/observation/terminalCollector.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 
 import { LogCategory, logger } from '@extension/services/loggingService';
-import type { RecorderLifecycleState } from '@extension/services/telemetry/recording/lifecycle/recorderLifecycleState';
+import type { RecorderLifecycleState } from '@extension/services/telemetry/recording/lifecycleController';
 import type { RecordedEvent } from '@extension/services/telemetry/recording/types';
 
 interface PendingExecution {
